+++ @fortawesome/free-brands-svg-icons/faVimeo.d.ts
@@ -1,4 +1,4 @@
-import { IconDefinition, IconPrefix, IconName } from "@fortawesome/fontawesome-common-types";
+import { IconDefinition, IconPrefix, IconName } from "@minddocdev/fontawesome-common-types";
 export const definition: IconDefinition;
 export const faVimeo: IconDefinition;
 export const prefix: IconPrefix;