+++ @fortawesome/pro-light-svg-icons/faScrubber.d.ts
@@ -1,4 +1,4 @@
-import { IconDefinition, IconPrefix, IconName } from "@fortawesome/fontawesome-common-types";
+import { IconDefinition, IconPrefix, IconName } from "@minddocdev/fontawesome-common-types";
 export const definition: IconDefinition;
 export const faScrubber: IconDefinition;
 export const prefix: IconPrefix;