+++ @fortawesome/pro-solid-svg-icons/faMp3Player.d.ts
@@ -1,4 +1,4 @@
-import { IconDefinition, IconPrefix, IconName } from "@fortawesome/fontawesome-common-types";
+import { IconDefinition, IconPrefix, IconName } from "@minddocdev/fontawesome-common-types";
 export const definition: IconDefinition;
 export const faMp3Player: IconDefinition;
 export const prefix: IconPrefix;